+++ app/code/Magento/Page/Block/Html/Head/Css.php
@@ -33,7 +33,7 @@ class Css extends \Magento\Core\Block\AbstractBlock
         $this->setAsset(
             $viewFileFactory->create(array(
                 'file' => (string)$this->getFile(),
-                'contentType' => \Magento\Core\Model\View\Publisher::CONTENT_TYPE_CSS
+                'contentType' => \Magento\View\Publisher::CONTENT_TYPE_CSS
             ))
         );
     }