+++ dev/tests/static/framework/Magento/TestFramework/Integrity/Library/PhpParser/StaticCalls.php
@@ -1,0 +1,104 @@
+<?php
+/**
+ * {license_notice}
+ *
+ * @copyright   {copyright}
+ * @license     {license_link}
+ */
+
+namespace Magento\TestFramework\Integrity\Library\PhpParser;
+
+/**
+ * Parse static calls and collect dependencies for it
+ *
+ * @package Magento\TestFramework
+ */
+class StaticCalls implements Parser, DependenciesCollector
+{
+    /**
+     * @var Tokens
+     */
+    protected $tokens;
+
+    /**
+     * Save static calls token key
+     *
+     * @var array
+     */
+    protected $staticCalls = array();
+
+    /**
+     * Collect dependencies
+     *
+     * @var array
+     */
+    protected $dependencies = array();
+
+    /**
+     * @param Tokens $tokens
+     */
+    public function __construct(Tokens $tokens)
+    {
+        $this->tokens = $tokens;
+    }
+
+    /**
+     * Check if it's foreign dependency
+     *
+     * @param array $token
+     * @return bool
+     */
+    protected function isTokenClass($token)
+    {
+        return is_array($token)
+            && !(in_array($token[1], array('static', 'self', 'parent')) || preg_match('#^\$#', $token[1]));
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function parse($token, $key)
+    {
+        if (is_array($token)
+            && $token[0] == T_PAAMAYIM_NEKUDOTAYIM
+            && $this->isTokenClass($this->tokens->getPreviousToken($key))
+        ) {
+            $this->staticCalls[] = $key;
+        }
+    }
+
+    /**
+     * Return class name from token
+     *
+     * @param int $staticCall
+     * @return string
+     */
+    protected function getClassByStaticCall($staticCall)
+    {
+        $step = 1;
+        $staticClassParts = array();
+        while ($this->tokens->getTokenCodeByKey($staticCall-$step) == T_STRING
+            || $this->tokens->getTokenCodeByKey($staticCall-$step) == T_NS_SEPARATOR
+        ) {
+            $staticClassParts[] = $this->tokens->getTokenValueByKey($staticCall-$step);
+            $step++;
+        }
+        return implode(array_reverse($staticClassParts));
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getDependencies(Uses $uses)
+    {
+        foreach ($this->staticCalls as $staticCall) {
+            $class = $this->getClassByStaticCall($staticCall);
+            if ($uses->hasUses()) {
+                $class = $uses->getClassNameWithNamespace($class);
+            }
+            $this->dependencies[] = $class;
+        }
+
+        return $this->dependencies;
+    }
+}