+++ dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_properties.php
@@ -99,4 +99,7 @@ return array(
     array('_backendFrontName', 'Magento\Backend\Helper\Data'),
     array('_app', 'Magento\Backend\Block\Widget\Grid\Column\Renderer\Currency'),
     array('_enginePool', '\Magento\Core\Block\Template\Context', '_engineFactory'),
+    array('types', '\Magento\Core\Model\Theme'),
+    array('_collectionFactory', '\Magento\Install\Controller\Action', 'themeProvider'),
+    array('_collectionFactory', '\Magento\Theme\Model\Config\Customization', 'themeProvider'),
 );