+++ lib/Magento/View/Design/Theme/Customization/FileAssetInterface.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * {license_notice}
+ *
+ * @category    Magento
+ * @package     Magento_Core
+ * @copyright   {copyright}
+ * @license     {license_link}
+ */
+
+/**
+ * Theme asset file interface
+ */
+namespace Magento\View\Design\Theme\Customization;
+
+interface FileAssetInterface
+{
+    /**
+     * Get content type of file
+     *
+     * @return string
+     */
+    public function getContentType();
+}