+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 * Various improvements:
+  * Added static test for check incorrect dependencies in library
+  * Moved Magento\Core\Model\Theme to Magento\View component
+  * Moved Magento\Core\Model\Design to Magento\View component
   * Consistent declaration of page-types
 * Fixed bugs:
   * Fixed unable to print order, invoice, or creditmemo in the front-end